[Vosk] Show download icon instead of downloading model without consent
When the app is opened and the model has not yet been downloaded, now a download icon is shown instead of starting the downlaod directly. On the first app start, for example, the user might be on mobile data or he might just want to disable Vosk right away and not do any download.